feat(location): added functionality for editing a user's location and cleaned up a bit
